Amended solution to use ContentTypeId
Corrected recursive method
Added clean app.config
