+++ Solutions/Governance.ContentTypeEnforceRetention/Governance.ContentTypeEnforceRetention.sln
@@ -1,9 +1,14 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
+VisualStudioVersion = 12.0.30110.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Governance.ContentTypeEnforceRetention", "Governance.ContentTypeEnforceRetention\Governance.ContentTypeEnforceRetention.csproj", "{27766626-6016-42A8-858B-EE4EC50C73BE}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Documentation", "Documentation", "{6089FE44-9C8F-4A8C-A3F7-BA5DCFD4AC78}"
+	ProjectSection(SolutionItems) = preProject
+		readme.md = readme.md
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
