+++ Solutions/Governance.ContentTypeEnforceRetention/Governance.ContentTypeEnforceRetention/App.config
@@ -1,0 +1,28 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <configSections>
+    <section name="Sites" type="System.Configuration.NameValueSectionHandler"/>
+    <section name="ContentTypeRetentionPolicyPeriod" type="System.Configuration.NameValueSectionHandler"/>
+  </configSections>
+
+  <startup>
+    <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5" />
+  </startup>
+
+  <appSettings>
+    <!-- The client id and client secret of the app as provided in /_layouts/15/appregnew.aspx -->
+    <add key="ClientID" value="enter id" />
+    <add key="ClientSecret" value="enter secret" />
+  </appSettings>
+
+  <Sites> <!--Site Collections to scan-->
+    <add key="site1" value="https://onebitdev2015.sharepoint.com/sites/contoso"/>
+    <add key="site2" value="https://onebitdev2015.sharepoint.com/sites/contosobeta"/>
+  </Sites>
+
+  <ContentTypeRetentionPolicyPeriod>
+    <!--Key is the Content Type ID and Value is the days of validity-->
+    <add key="0x0101009148F5A04DDD49cbA7127AADA5FB792B006973ACD696DC4858A76371B2FB2F439A" value="183" /> <!--Audio-->
+    <add key="0x0101" value="365" /> <!--Document-->
+  </ContentTypeRetentionPolicyPeriod>
+</configuration>
