+++ Solutions/Governance.ContentTypeEnforceRetention/Governance.ContentTypeEnforceRetention/Properties/AssemblyInfo.cs
